api/requirements.txt
click==8.0.2
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2-binary
PyJWT==2.2.0
six==1.16.0
uWSGI==2.0.20
Werkzeug==2.0.2
